:recycle: Move StockChartService to own file, add new hook
